single instance music player with proper beat disposition and aimed delegate callbacks

+++ Blast Cipher/Assets/Code/Standard/Scriptable Objects/Containers/AudioContainer.cs
@@ -4,5 +4,5 @@
 public class AudioContainer : ScriptableObject
 {
 	public AudioClip[] tracks;
-	public int[] bpmValues;
+	public float[] bpmValues;
 }
